# api/docs/swagger.yaml
definitions:
  api.AchievementGroupResponse:
    properties:
      active:
        example: true
        type: boolean
      description:
        example: Достижения в научной деятельности
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Научная деятельность
        type: string
    required:
    - active
    - description
    - id
    - name
    type: object
...
  api.CreateAchievementGroupRequest:
    properties:
      description:
        example: Достижения в научной деятельности
        type: string
      name:
        example: Научная деятельность
        type: string
    required:
    - description
    - name
    type: object
...
  api.CredentialsRequest:
    properties:
      password:
        example: secret123
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    - username
    type: object
...
  api.ServerError:
    properties:
      code:
        example: SERVER_ERROR
        type: string
      details:
        type: string
      message:
        example: Internal server error
        type: string
      ruMessage:
        example: Внутренняя ошибка сервера
        type: string
    type: object
...
  api.UnauthorizedError:
    properties:
      code:
        example: UNAUTHORIZED
        type: string
      details:
        type: string
      message:
        example: Unauthorized access
        type: string
      ruMessage:
        example: Неавторизованный доступ
        type: string
    type: object
...
info:
  contact: {}
  description: API for managing SESC departments, users and permissions
  title: SESC Management API
  version: "1.0"
paths:
  /achievement-groups:
    get:
      description: Retrieves all achievement groups
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - default: false
        description: Show inactive groups
        in: query
        name: show_inactive
        type: boolean
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.AchievementGroupResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Get all achievement groups
      tags:
      - achievement-groups
    post:
      consumes:
      - application/json
      description: Creates a new achievement group
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: Group details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateAchievementGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.AchievementGroupResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api.InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Create new achievement group
      tags:
      - achievement-groups
  /achievement-groups/{id}:
    patch:
      consumes:
      - application/json
      description: Updates an achievement group
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      - description: Group fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.PatchAchievementGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AchievementGroupResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api.InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/api.GroupNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Update achievement group
      tags:
      - achievement-groups
  /achievement-templates:
    get:
      description: Retrieves all achievement templates
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - default: false
        description: Show inactive templates
        in: query
        name: show_inactive
        type: boolean
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.AchievementTemplateResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Get all achievement templates
      tags:
      - achievement-templates
    post:
      consumes:
      - application/json
      description: Creates a new achievement template
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: Template details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateAchievementTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.AchievementTemplateResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api.InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/api.GroupNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Create new achievement template
      tags:
      - achievement-templates
  /achievement-templates/{id}:
    patch:
      consumes:
      - application/json
      description: Updates an achievement template
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: Template UUID
        in: path
        name: id
        required: true
        type: string
      - description: Template fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.PatchAchievementTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AchievementTemplateResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api.InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/api.AchievementTemplateNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Update achievement template
      tags:
      - achievement-templates
  /auth/admin/login:
    post:
      consumes:
      - application/json
      description: Verifies admin token and returns a JWT token with admin privileges
      parameters:
      - description: Admin credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CredentialsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api.InvalidRequestError'
        "401":
          description: Invalid admin token or not recognized
          schema:
            $ref: '#/definitions/api.CredentialsNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: Admin login
      tags:
      - authentication
  /auth/credentials/{id}:
    delete:
      description: Deletes credentials for a user
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/api.InvalidUUIDError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: User credentials not found
          schema:
            $ref: '#/definitions/api.CredentialsNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Delete user credentials
      tags:
      - authentication
    get:
      description: Retrieves credentials for a user
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CredentialsRequest'
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/api.InvalidUUIDError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: User not found or does not exist
          schema:
            $ref: '#/definitions/api.CredentialsNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Get user credentials
      tags:
      - authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifies user credentials and returns a JWT token
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CredentialsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api.InvalidRequestError'
        "401":
          description: Invalid credentials or user does not exist
          schema:
            $ref: '#/definitions/api.CredentialsNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: User login
      tags:
      - authentication
  /auth/validate:
    get:
      description: Validates a JWT token and returns the identity information
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.IdentityResponse'
        "401":
          description: Invalid token
          schema:
            $ref: '#/definitions/api.InvalidTokenError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Validate JWT token
      tags:
      - authentication
  /departments:
    get:
      description: Retrieves list of all registered departments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DepartmentsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: List all departments
      tags:
      - departments
    post:
      consumes:
      - application/json
      description: Creates a new department with the given details
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: Department details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Department'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api.InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "409":
          description: Department with this name already exists
          schema:
            $ref: '#/definitions/api.DepartmentExistsError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Create a new department
      tags:
      - departments
  /departments/{id}:
    delete:
      description: Deletes a department by its ID
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: Department UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/api.InvalidDepartmentIDError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: Department not found
          schema:
            $ref: '#/definitions/api.DepartmentNotFoundError'
        "409":
          description: Cannot remove department, it still has some users
          schema:
            $ref: '#/definitions/api.CannotRemoveDepartmentError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Delete a department
      tags:
      - departments
    put:
      consumes:
      - application/json
      description: Updates an existing department with new details
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: Department UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated department details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Department'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/api.InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: Department not found
          schema:
            $ref: '#/definitions/api.DepartmentNotFoundError'
        "409":
          description: Department with this name already exists
          schema:
            $ref: '#/definitions/api.DepartmentExistsError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Update department details
      tags:
      - departments
  /files:
    get:
      consumes:
      - application/json
      description: Returns a list of files based on search criteria
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: File name to search for
        in: query
        name: name
        type: string
      - description: Owner ID to filter by
        in: query
        name: owner_id
        type: string
      - description: If true, return only common files
        in: query
        name: common
        type: boolean
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      - default: 50
        description: Pagination limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.FileListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - BearerAuth: []
      summary: Search files
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      description: Uploads a new file. Admin users create common files, regular users
        create files owned by themselves.
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.FileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - BearerAuth: []
      summary: Upload a file
      tags:
      - files
  /files/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a file by ID
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - BearerAuth: []
      summary: Delete file
      tags:
      - files
...
  /roles:
    get:
      description: Retrieves all system roles with their permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RolesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.Error'
      summary: List all roles
      tags:
      - roles
  /users:
    get:
      description: Retrieves detailed information about all users
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Get all users registered in the system
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user with specified role (non-teacher)
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Invalid name specified
          schema:
            $ref: '#/definitions/api.InvalidNameError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Create new user
      tags:
      - users
  /users/{id}:
    get:
      description: Retrieves detailed information about a user
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/api.InvalidUUIDError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/api.UserNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Get user details
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Applies a partial update to the user identified by {id}. Only non-nil
        fields in the request are applied.
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: User fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Invalid name
          schema:
            $ref: '#/definitions/api.InvalidNameError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/api.UserNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Partially update user
      tags:
      - users
  /users/{id}/credentials:
    put:
      consumes:
      - application/json
      description: Assigns username/password credentials to an existing user
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CredentialsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: AuthID
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid credentials format
          schema:
            $ref: '#/definitions/api.InvalidCredentialsError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/api.ForbiddenError'
        "404":
          description: User does not exist
          schema:
            $ref: '#/definitions/api.UserNotFoundError'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/api.UserExistsError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Register user credentials
      tags:
      - authentication
  /users/me:
    get:
      description: Returns information about the current authenticated user
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "401":
          description: Unauthorized or invalid token
          schema:
            $ref: '#/definitions/api.UnauthorizedError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/api.UserNotFoundError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ServerError'
      security:
      - BearerAuth: []
      summary: Get current user information
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Enter 'Bearer ' followed by your token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"